Finish water detection, improve grass detection
This still has a little room left for fine tuning, and it's still written
specifically for FRLG, but overall the script has improved a lot. It detects
water now by looking for certain colors instead of comparing each block in the
image to each of the sprites we're looking for.

What I want to do with this is use the database table models to store the data
as the script runs. If the --commit flag was present, they'll just be added at
that point. This is my next goal before I finally add all of FRLG to the
database. I also want to add a check before any database commits happen to make
sure that there aren't any stored patches for that route. It seems like I would
forget about it at some point and end up with some duplicates.