Woo, able to process all 25 Kanto routes
I reached most (all?) of the goals of my last commit. Now I just need to add the
other games. Considering packing this script on its own, because clearly there
needs to be some way to store the individual color values for each game. I'm
sure it isn't even the same between FRLG and RSE.

On a bright note, it only took 1m11s to add routes 1-25 of Kanto (gen3) to the
database.