Updates to routesearch for RSE, create insertworld
I updated the image searcher to look for tall grass in RSE, but I think I need
to change things around. My plan is to use a list of lists, with each inner list
representing a type of grass to check for, since there are more than two colors
of grass in RSE (the ash-covered grass isn't currently detected, though the tall
grass and the taller grass are now).

Also, insertworld.py just reads a csv file and inserts it into a database. It's
for image mapping a world map.